requirements.txt
streamlit
pandas
os
numpy
matplotlib.pyplot
seaborn
plotly.express